feat(config): provider types
Providers have a new field called 'type', which is a way to easily
populate commonly shared fields of different git forges such as API
paths for the different OAuth things.

Current supported values are GitHub, GitLab, Forgejo and Gitea, not case
sensitive. (Gitea is just an alias for the Forgejo defaults as they are
currently the same)